requirements.txt
fastapi
uvicorn
transformers
pydantic
torch
torchvision 
pillow
python-multipart